requirements.txt
streamlit
pillow
PyPDF2
python-docx